pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "music-bingo"
version = "0.1.0"
description = "Offline music bingo card generator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "Flask>=3.0.0",
  "Pillow>=9.1.0",
  "qrcode[pil]>=7.4.2",
  "python-dotenv>=1.0.0",
  "reportlab>=4.0.0",
  "spotipy>=2.23.0",
  "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["music_bingo*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
